Moved GRPCStream to connections directory
